pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "mucus"
authors = [{name = "Chris Coleman", email = "chris@chriscoleman.uk"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = ["appdirs", "click", "confuse", "redis", "rich", "myass", "pycryptodome"] 

[project.scripts]
mucus = "mucus.cli:main"
